backend/pdfcollab/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.3.0
django-filter==23.5